squash a bunch of compiler warnings - some of the remaining signed/unsigned warning might take a little more thought to fix

+++ cyrus/imap/ctl_deliver.c
@@ -1,5 +1,5 @@
 /* ctl_deliver.c -- Program to perform operations on duplicate delivery db
- * $Id: ctl_deliver.c,v 1.22 2007/03/30 18:40:20 murch Exp $
+ * $Id: ctl_deliver.c,v 1.23 2007/09/28 02:27:46 murch Exp $
  * Copyright (c) 1998-2003 Carnegie Mellon University.  All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -124,7 +124,7 @@ int main(int argc, char *argv[])
 		"using cyr_expire -E instead\n");
 
 	r = snprintf(buf, sizeof(buf), "%s/cyr_expire", SERVICE_PATH);
-	if(r < 0 || r >= sizeof(buf)) {
+	if(r < 0 || r >= (int) sizeof(buf)) {
 	    fatal("cyr_expire command buffer not sufficiently big", EC_CONFIG);
 	}
